Finish client functions

+++ TechAssist/src/main/java/com/example/techassist/DTO/AccountBalanceDTO.java
@@ -1,9 +1,10 @@
 package com.example.techassist.DTO;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class AccountBalanceDTO {
     public List<CallDTO> callDTOList;
-    public float serviceFee;
+    public BigDecimal serviceFee;
 }
